requirements_training.txt
easing_functions==1.0.4
tensorboard
torch==2.2.2
torchvision==0.17.2
tqdm==4.61.1
h5py
opencv-python
gin-config
prefetch_generator